disable "previous" button on welcome page
It is not clear from GTK docs, but it looks like switch-page cb would never
be called for an empty notebook. In addition, switch-change cb is called
from gtk_widget_show_all and it seems there gtk_notebook_get_current_page
always returns -1. The effect of all that is that we should not check if
notebook is empty in switch-page cb, otherwise "previous" button will
always be enabled on "welcome" page when it is shown for the first time.

Simplify the code a bit while at it.

Closes https://github.com/maemo-leste/bugtracker/issues/168

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>